Add atoms support
Currently we have thousands of duplicated strings due QString usage. E.g. 100 lights with the same manufacturer name easily result in 100 times "Ikea of Sweden". In case of QStrings they are allocated on the heap as UTF-16 which is 2 bytes per character. Atoms can also be strings but are only allocated once.
Each atom has a unique index which stays stable for the life time of the application. In contrast to normal strings comparing atoms is very fast since we only compare integers.
They will be used by other classes in future and mostly replace QStrings.